requirements.txt
# Image processing
Pillow>=10.0.0
imagehash>=4.3.1
opencv-python>=4.8.0

# Face detection and recognition
# Note: dlib compilation requires python3.11-dev system package
# Alternative: use 'pip install dlib-binary' for pre-built version
face_recognition>=1.3.0
face_recognition_models>=0.3.0
dlib>=19.24.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
click>=8.1.0
